Create frames
there are 3 frames first frame is animal and bird hyper link are there in a first frame .click the animal, bird link  and visible the list of animals and  birds are visible in a second a frame. for example second frame in list of animal and birds are there and click the lion and parrot visible the image in a third frame.